requirements.txt
streamlit
ibm-watsonx-ai
python-dotenv
PyPDF2
graphviz
